pyproject.toml
[project]
name = "deepagents"
version = "0.0.11"
description = "General purpose 'deep agent' with sub-agent spawning, todo list capabilities, and mock file system. Built on LangGraph."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11,<4.0"
dependencies = [
    "langchain>=1.0.0,<2.0.0",
    "langchain-core>=1.0.0,<2.0.0",
    "langchain-anthropic>=1.0.0,<2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "build",
    "twine",
]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["deepagents", "tests"]
[tool.setuptools.package-dir]
"deepagents" = "src/deepagents"
"tests" = "tests"

[tool.setuptools.package-data]
"*" = ["py.typed"]

[dependency-groups]
dev = [
    "pyright>=1.1.406",
    "python-dotenv>=1.1.1",
]
